[api] Prohibit certain VMs from joining server groups
We cannot adhere to affinity-rules if we handle big VMs differently on
the cluster or when spawning smaller VMs on specifically prepared hosts.
We could just ignore that, but we want the user to know about it.
Therefore, we prohibit big VMs from being created, if the
creation-request contains a server group.