Implement freeing a host for spawning big VMs
We need to be able to spawn big VM and thus have to prepare a host for
it i.e.  free it from other VMs. This cannot be mapped in nova
currently, as nova has no view on the individual hosts in a VMware
cluster. Therefore, we implement some function in the VMware driver to
actually free up the host and also release it into the cluster again.
Placement is used for keeping a view over all clusters, as we only need
one host free per HV size. The conductor is responsible for managing
this stuff, as we need a service talking to all compute nodes and
additionally to placement. It seemed like the best choice other than
starting a new service for this.